refactor: rename statusapi pkg to httpapi
